Homework_3/requirements.txt
langchain>=0.3.0
langgraph>=0.2.27
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-tavily>=0.1.0
tavily-python>=0.5.0
wikipedia>=1.4.0
wolframalpha>=5.0.0
python-dotenv>=1.0.1